+++ packages/core/src/app/shipping/ConsignmentAddressSelector.tsx
@@ -9,6 +9,7 @@ import {
     type AddressFormValues,
     AddressSelect,
     AddressType,
+    decodeAddressLabel,
     isValidAddress,
     mapAddressFromFormValues,
 } from '../address';
@@ -42,7 +43,7 @@ const ConsignmentAddressSelector = ({
     const [createCustomerAddressError, setCreateCustomerAddressError] = useState<Error>();
 
     const {
-        userJourney: { hasCompanyAddressBook },
+        userJourney: { hasCompanyAddressBook, hasAddressLabel },
     } = useCapabilities();
 
     const {
@@ -51,17 +52,22 @@ const ConsignmentAddressSelector = ({
         updateConsignment,
         createCustomerAddress,
         customer,
-        validateMaxLength,
         getConsignments: getPreviousConsignments,
     } = useShipping();
 
     // TODO: add filter for addresses
-    const addresses = customer.addresses || EMPTY_ARRAY;
+    const addresses = (customer.addresses || EMPTY_ARRAY).map((address) =>
+        decodeAddressLabel(address, hasAddressLabel),
+    );
+    const decodedSelectedAddress =
+        selectedAddress && decodeAddressLabel(selectedAddress, hasAddressLabel);
 
     const isGuest = customer.isGuest;
 
-    const handleSelectAddress = async (address: Address) => {
-        if (!isValidAddress(address, getFields(address.countryCode), validateMaxLength)) {
+    const handleSelectAddress = async (rawAddress: Address) => {
+        const address = decodeAddressLabel(rawAddress, hasAddressLabel);
+
+        if (!isValidAddress(address, getFields(address.countryCode), true)) {
             return onUnhandledError(new AssignItemInvalidAddressError());
         }
 
@@ -155,13 +161,13 @@ const ConsignmentAddressSelector = ({
                 isOpen={isOpenNewAddressModal}
                 onRequestClose={handleCloseAddAddressForm}
                 onSaveAddress={handleSaveAddress}
-                selectedAddress={isGuest ? selectedAddress : undefined}
+                selectedAddress={isGuest ? decodedSelectedAddress : undefined}
                 shouldShowSaveAddress={hasCompanyAddressBook}
             />
             {isGuest ? (
                 <GuestCustomerAddressSelector
                     onUseNewAddress={handleUseNewAddress}
-                    selectedAddress={selectedAddress}
+                    selectedAddress={decodedSelectedAddress}
                 />
             ) : (
                 <AddressSelect
@@ -169,7 +175,7 @@ const ConsignmentAddressSelector = ({
                     onSelectAddress={handleSelectAddress}
                     onUseNewAddress={handleUseNewAddress}
                     placeholderText={<TranslatedString id="shipping.choose_shipping_address" />}
-                    selectedAddress={selectedAddress}
+                    selectedAddress={decodedSelectedAddress}
                     showSingleLineAddress
                     type={AddressType.Shipping}
                 />