+++ packages/utility/src/index.ts
@@ -1,4 +1,5 @@
 export { default as isBuyNowCart } from './isBuyNowCart';
+export { default as isExperimentEnabled } from './isExperimentEnabled';
 export { default as navigateToOrderConfirmation } from './navigateToOrderConfirmation';
 export { default as isErrorWithTranslationKey } from './is-error-with-translation-key';
 export { default as hideEditCartLink } from './hideEditCartLink';
@@ -7,3 +8,4 @@ export {
     type B2BStoredAddressIds,
     type B2BStoredPaymentValues,
 } from './storage/B2BSessionStorage';
+export { CannotCreatePersonalAccountSessionStorage } from './storage/CannotCreatePersonalAccountSessionStorage';